VoiceOrder-Backend/requirements.txt
fastapi
uvicorn
python-dotenv
googlemaps
requests
pydantic
